refactor: consolidate agent modules
