+++ conversation_service/agents/__init__.py
@@ -5,10 +5,10 @@ conversation service so they can be imported directly from
 ``conversation_service.agents``.
 """
 
-from .entity_extractor_agent import EntityExtractionCache, EntityExtractorAgent
-from .intent_classifier_agent import IntentClassificationCache, IntentClassifierAgent
-from .query_generator_agent import QueryGeneratorAgent, QueryOptimizer
-from .response_generator_agent import ResponseGeneratorAgent
+from .entity_extractor import EntityExtractionCache, EntityExtractorAgent
+from .intent_classifier import IntentClassificationCache, IntentClassifierAgent
+from .query_generator import QueryGeneratorAgent, QueryOptimizer
+from .response_generator import ResponseGeneratorAgent, stream_response
 
 __all__ = [
     "IntentClassifierAgent",
@@ -18,5 +18,6 @@ __all__ = [
     "QueryGeneratorAgent",
     "QueryOptimizer",
     "ResponseGeneratorAgent",
+    "stream_response",
 ]
